refactor: remove expandedAssetSheet and move chart files to /value-chart
