--- src/components/expanded-state/chart/index.js
@@ -1,1 +1,0 @@
-export { default as ChartExpandedStateHeader } from './ChartExpandedStateHeader';
