+++ src/components/expanded-state/index.js
@@ -1,3 +1,2 @@
-export { default as ChartExpandedState } from './ChartExpandedState';
 export { default as ContactProfileState } from './ContactProfileState';
 export { default as WalletProfileState } from './WalletProfileState';
