+++ src/components/value-chart/chart-data-labels/ChartHeaderSubtitle.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { TruncatedText } from '../../../text';
+import { TruncatedText } from '@rainbow-me/components/text';
 
 const ChartHeaderSubtitle = styled(TruncatedText).attrs(
   ({
