+++ src/components/value-chart/chart-data-labels/index.js
